Add hashing interface to merkledb
